% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace.R
\name{estforboot}
\alias{estforboot}
\title{estforboot (estforboot)}
\usage{
estforboot(X, Z, S, Y, family.Y)
}
\arguments{
\item{X}{is a matrix of pre-treatment covariates without intercept  (n x p).}

\item{Z}{is a vector of treatment  (n x 1).}

\item{S}{is a vector of binary intermidiate outcome  (n x 1).}

\item{Y}{is a vector of outcome  ( n x 1).}

\item{family.Y}{specifies the family for the outcome model.
\code{"gaussian"}: a linear regression model for the continuous outcome.

\code{"binomial"}: a logistic regression model for the binary outcome.}
}
\value{
all point estimators
}
\description{
A function calculates all estimators
}
