% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-PACEFit.R
\name{neighbourSlopes}
\alias{neighbourSlopes}
\alias{neighbourSlopes,PACEFit-method}
\title{Shrunken neighbour proximity slopes}
\usage{
neighbourSlopes(object, ...)

\S4method{neighbourSlopes}{PACEFit}(object, ...)
}
\arguments{
\item{object}{A \link{PACEFit}.}

\item{...}{Unused.}
}
\value{
A data frame of shrunken slopes.
}
\description{
Returns the shrunken per-(gene, focal, neighbour) proximity coefficients with
their local false sign rate (lfsr). For condition cohorts the responder
interaction terms are included.
}
\examples{
\dontrun{ neighbourSlopes(fit) }
}
