% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pace.r
\name{plot_pace}
\alias{plot_pace}
\title{Plot PACE Forecast vs Actuals}
\usage{
plot_pace(forecast_df, actual_df = NULL)
}
\arguments{
\item{forecast_df}{Data frame from pace_forecast()}

\item{actual_df}{Original historical data (optional)}
}
\value{
ggplot object
}
\description{
Plot PACE Forecast vs Actuals
}
\examples{
df <- data.frame(Time = seq(as.Date("2020-01-01"), by="quarter", length.out=12),
                 Value = c(80,95,100,92,105,110,90,78,70,85,98,110))
fc <- pace_forecast(df, 4)
plot_pace(fc, df)
}
