% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/paceFit.R
\name{paceShrink}
\alias{paceShrink}
\alias{paceShrink,PACEFit-method}
\title{Shrink the neighbour slopes}
\usage{
paceShrink(object, ...)

\S4method{paceShrink}{PACEFit}(object, ...)
}
\arguments{
\item{object}{A \link{PACEFit} from \code{\link[=paceModel]{paceModel()}}.}

\item{...}{Unused.}
}
\value{
The \code{PACEFit} with the shrunken neighbour slopes added.
}
\description{
Stabilises the fitted per-(gene, focal, neighbour) proximity slopes with
multivariate adaptive shrinkage, populating \code{\link[=neighbourSlopes]{neighbourSlopes()}}.
}
\examples{
fit <- readRDS(system.file("extdata", "pace_fit_example.rds", package = "PACE"))
fit <- paceShrink(fit)
head(neighbourSlopes(fit))
}
