% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-PACEFit.R
\name{topDrivers}
\alias{topDrivers}
\alias{topDrivers,PACEFit-method}
\title{Per-pair driver tables}
\usage{
topDrivers(object, ...)

\S4method{topDrivers}{PACEFit}(object, ...)
}
\arguments{
\item{object}{A \link{PACEFit}.}

\item{...}{Unused.}
}
\value{
A list of per-pair driver tables.
}
\description{
Returns the driver-score tables ranking the genes that mediate each
focal-neighbour spatial relationship.
}
\examples{
fit <- readRDS(system.file("extdata", "pace_fit_example.rds", package = "PACE"))
names(topDrivers(fit))
}
